refactor(MP-34): Start folders names (connected with table) with capitalize letter.
