+++ src/components/Table/ImageCell.tsx
